feat: Adds connection button to GUI
Also renames the App.py file to be more descriptive of the script's
purpose. The button is not in its final position and there is no way to
connect twice. This is because opening the GoPro starts a thread and
closing it does not end the thread. Reconnecting to the GoPro would also
require touching the GoPro again so there is no reason to need the
button to work twice when the app could just be opened again.

Closes #1 Issue